correct all view html code add authorize without role(admin)

+++ CarStoreWeb/Controllers/DeclarationController.cs
@@ -4,23 +4,28 @@ using System.Linq;
 using System.Threading.Tasks;
 using CarStoreWeb.Models;
 using CarStoreWeb.Models.ViewModels;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 
 namespace CarStoreWeb.Controllers
 {
+    [Authorize]
     public class DeclarationController : Controller
     {
+        private IAuthorizationService _authService;
         ICarRepository _repository;
         const int _pageSize = 3;
-        public DeclarationController(ICarRepository repo)
+        public DeclarationController(ICarRepository repo, IAuthorizationService auth)
         {
             _repository = repo;
+            _authService = auth;
         }
+        [AllowAnonymous]
         public IActionResult List(string category, int pageNum, SortState? sortOrder)
         {
             var cars = _repository.Cars
-                    .Where(p => category == null || p.Brand == category);
+                    .Where(c => category == null || c.Brand == category);
 
             switch (sortOrder)
             {
@@ -56,7 +61,7 @@ namespace CarStoreWeb.Controllers
                     .Take(_pageSize);
             return View(new DeclarationListViewModel()
             {
-                Cars = cars,
+                Cars=cars,// = cars.Select(c => { c.ThisAuthor = HttpContext.User.Identity.Name == c.Author; return c; }),
                 CurrentCategory = category,
                 PagingInfo = new PagingInfo()
                 {
@@ -65,43 +70,77 @@ namespace CarStoreWeb.Controllers
                     TotalItems = category == null ? _repository.Cars.Count() :
                     _repository.Cars.Count(c => c.Brand == category)
                 },
-                SortViewModel = new SortViewModel(sortOrder)
+                SortViewModel = new SortViewModel(sortOrder),
+                IsAuthenticated = HttpContext.User.Identity.IsAuthenticated
             });
         }
+        
         [HttpGet]
         public IActionResult AddItem()
         {
-            return View(new Car());
+            return View("EditItem",new Car()); 
         }
 
-        [HttpPost]
-        public IActionResult AddItem(Car car)
-        {
-            if (ModelState.IsValid)
-            {
-                _repository.AddCar(car);
-                return RedirectToAction("Completed");
-            }
-            else
-                return View(car);
-        }
-
-        [HttpPost]
-        public IActionResult RemoveItem(int CarID, string returnUrl)
+        [HttpGet]
+        public async Task<IActionResult> EditItem(int CarID)
         {
             var car = _repository.FindCar(CarID);
+            var authorized = await _authService.AuthorizeAsync(User, car, "Authors");
 
-            if (car != null)
-            {
-                _repository.DeleteCar(CarID);
-            }
+            if (!authorized.Succeeded)
+                return StatusCode(403);
 
-            return LocalRedirect(returnUrl); //its safety
+            return View(_repository.FindCar(CarID));
         }
 
-        public ViewResult Completed()
+
+        [HttpPost]
+        public async Task<IActionResult> EditItem(Car car)
         {
-            return View();
+            if (!ModelState.IsValid)
+                return View(car);            
+
+            if (car.CarID == 0)
+            {
+                car.Author = HttpContext.User.Identity.Name;
+                car=_repository.AddCar(car);
+                TempData["message"] = $"{car.Brand} {car.Model} has been saved!";
+                return RedirectToAction("Completed", new { CarID = car.CarID });
+            }
+            else  //(car.CarID != 0)
+            {
+                var authorized = await _authService.AuthorizeAsync(User, car, "Authors");
+
+                if (!authorized.Succeeded)
+                    return StatusCode(403);
+
+                _repository.EditCar(car);
+                TempData["message"] = $"{car.Brand} {car.Model} has been edited!";
+                return RedirectToAction("Completed",new {CarID= car.CarID });
+            }
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> RemoveItem(int CarID)
+        {
+            var car = _repository.FindCar(CarID);
+            var authorized = await _authService.AuthorizeAsync(User, car, "Authors");
+
+            if (car != null && authorized.Succeeded)
+            {
+                _repository.DeleteCar(CarID);
+                TempData["message"] = $"{car.Brand} {car.Model} was deleted!";
+                return RedirectToAction("List"); //its safety
+            }
+            else
+                return StatusCode(403);
+
+        }
+
+        public ViewResult Completed(int CarID)
+        {
+            var car = _repository.FindCar(CarID);
+            return View(car);
         }
     }
 }
